feat: #253 FriendsSettings 컴포넌트와 라우트를 제거하고, ExploreFriends로 이동
